// src/types/types.ts
export interface Character {
	name: string;
	height: string;
	mass: string;
	hair_color: string;
	skin_color: string;
	eye_color: string;
	birth_year: string;
	gender: string;
	homeworld: string;
	films: string[];
	species: string[];
	vehicles: string[];
	starships: string[];
	url: string;
}

export interface ApiResponse<T> {
	count: number;
	next: string | null;
	previous: string | null;
	results: T[];
}